Update main.lua
Issue: when using progress bar it locked players into an animation and would not cancel.

Fix: 
Updated on finish to cancel current ped task. 
Updated the controls for canceling in the main loop.
Updated the disable control actions so that they disable the appropriate input(s),

Expected result: We should now be able to play an animation on the player ped and on cancel/complete return the player ped back to its default animation. When we right click or esc it should cancel the progressbar + animation. Finally when riding a horse, if specified to be disabled, player should no longer be able to move the horse